fixed backspace key action issue in Mac

+++ src/app/components/common/node-editor/node-editor.component.ts
@@ -191,7 +191,7 @@ export class NodeEditorComponent implements OnInit {
     if (event.keyCode == 32) {
       resetNodes(this.flowEditorService);
     }
-    if (event.key === 'Delete' || event.key == 'Backspace') {
+    if (event.key === 'Delete' || (event.key == 'Backspace' && event.metaKey)) {
       this.deleteSelectedEntity();
     }
   }
